+++ CHANGELOG.md
@@ -15,6 +15,7 @@ under development
 - Add `Context.QueryParams` and `Context.QueryParam`.
 - Add `Context.JSONP` and `Context.JSONPCallback`.
 - Add `Context.Render` to render a template.
+- Add `Context.RouteURL` to generate URL of the naming route.
 
 v1.8.1 April 2, 2020
 --------------------